requirements.txt
appdirs==1.4.4
APScheduler==3.9.1.post1
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.5
deepdiff==4.3.2
Django==1.11.7
dukpy==0.2.3
future==0.18.2
HTMLTestRunner-Python3==0.8.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itchat==1.2.32
javascripthon==0.12
Jinja2==3.0.3
jupyter-echarts-pypkg==0.1.2
lml==0.0.2
MarkupSafe==2.0.1
ordered-set==4.0.2
Pillow==8.4.0
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.2.0
pyee==8.2.2
PyMySQL==1.0.2
pypng==0.20220715.0
pyppeteer==0.2.6
PyQRCode==1.2.1
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==3.141.0
six==1.16.0
tqdm==4.64.1
typing_extensions==4.1.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
websockets==9.1
wxpy==0.3.9.8
xlrd==2.0.1
zipp==3.6.0
